message: Add message routers and handlers
Define message handlers for the render and browser processes (to enable
inter-process communication - see
https://bitbucket.org/chromiumembedded/cef/wiki/GeneralUsage#markdown-header-inter-process-communication-ipc).

In short this allows us to send messages from js to C++ in the main
"browser" process of cef. The two functions ("gstSendMsg" and
"gsCancelMsg") are attached to the javascript global "window" object.

Note that we need to add a RendererApp in this commit to contain and run
the javascript side of the message routing.

See cef sample for more detail:
https://github.com/chromiumembedded/cef-project/tree/aaccb78661c390a5ce6650df8db54ea5424d431c/examples/message_router